Mostly fixed UART example
Still not sure what the expected behavior is, but the app waits for a
JLink string from RTT and echoes it back over RTT. However, when
connecting to the app via BTLE using the NRF UART iPhone app, it
endlessly loops sending the most recent message. While this is easily
fixable, I will wait to fix it until I know what the app is supposed to
do with regards to its data source. Should it get its input from RTT, or
the iPhone app?